src/requirements.txt
click>=8.0.0
pyyaml>=6.0
pdfplumber>=0.11.8
pdfminer.six>=20251107
numpy>=1.24.0
httpx>=0.27.0
mcp>=1.0.0
openai>=1.0.0
pytest>=7.0.0
